requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
Django==4.1
efficientnet-pytorch==0.7.1
facenet-pytorch==2.5.2
fonttools==4.38.0
idna==3.3
imutils==0.5.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
mlxtend==0.21.0
nose==1.3.7
numpy==1.23.2
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-model-summary==0.1.2
pytube==12.1.0
pytz==2022.7
requests==2.28.1
resnest==0.0.5
scikit-learn==1.2.0
scipy==1.9.1
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
